Menu3

+++ team_jg_entertainment/src/main/java/lv/javaguru/java2/jg_entertainment/restaurant/core/services/services_menu/RemoveMenuService.java
@@ -2,7 +2,11 @@ package lv.javaguru.java2.jg_entertainment.restaurant.core.services.services_men
 
 import lv.javaguru.java2.jg_entertainment.restaurant.core.database.DatabaseMenu;
 import lv.javaguru.java2.jg_entertainment.restaurant.core.requests.menus.RemoveMenuRequest;
+import lv.javaguru.java2.jg_entertainment.restaurant.core.responses.menus.CoreError;
 import lv.javaguru.java2.jg_entertainment.restaurant.core.responses.menus.RemoveMenuResponse;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class RemoveMenuService {
 
@@ -13,6 +17,12 @@ public class RemoveMenuService {
     }
 
     public RemoveMenuResponse execute(RemoveMenuRequest request) {
+        if (request.getMenuNumberToRemove() == null) {
+            CoreError error = new CoreError("id", "Must not be empty!");
+            List<CoreError> errors = new ArrayList<>();
+            errors.add(error);
+            return new RemoveMenuResponse(errors);
+        }
         boolean isMenuRemoved = databaseMenu.deleteByNr(request.getMenuNumberToRemove());
         return new RemoveMenuResponse(isMenuRemoved);
     }
